requirements.txt
falcon
future
hidapi
jsonschema
tinydb